fix: improve profiling accuracy and update remove old requestRate

+++ libs/container_agent/profiler.h
@@ -50,8 +50,8 @@ public:
 
     struct sysStats {
         uint64_t timestamp = 0;
-        double cpuUsage = 0;
-        long memoryUsage = 0;
+        int cpuUsage = 0;
+        int memoryUsage = 0;
         unsigned int gpuUtilization = 0;
         unsigned int gpuMemoryUsage = 0;
         long maxGpuMemoryUsage = 0;
@@ -81,9 +81,9 @@ private:
 
     bool cleanupNVML();
 
-    double getCPUInfo(unsigned int pid);
+    int getCPUInfo(unsigned int pid);
 
-    long getMemoryInfo(unsigned int pid);
+    int getMemoryInfo(unsigned int pid);
 
     nvmlAccountingStats_t getGPUInfo(unsigned int pid, nvmlDevice_t device);
